Refactor : Edit class name
I edit user to member
