+++ src/main/java/econo/app/sleeper/service/member/MemberService.java
@@ -13,7 +13,7 @@ import java.util.Optional;
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
-public class UserService {
+public class MemberService {
 
     private final MemberRepository memberRepository;
