+++ src/main/java/econo/app/sleeper/web/member/MemberController.java
@@ -2,8 +2,8 @@ package econo.app.sleeper.web.member;
 
 import econo.app.sleeper.domain.member.Member;
 import econo.app.sleeper.service.character.CharacterService;
+import econo.app.sleeper.service.member.MemberService;
 import econo.app.sleeper.service.money.MoneyService;
-import econo.app.sleeper.service.member.UserService;
 import econo.app.sleeper.service.sleep.SleepAdvisorService;
 import econo.app.sleeper.web.character.dto.InitialCharacterDto;
 import econo.app.sleeper.web.common.CommonResponse;
@@ -23,8 +23,8 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 @RequiredArgsConstructor
 @Tag(name = "user", description = "사용자 관련 API")
-public class UserController {
-    private final UserService userService;
+public class MemberController {
+    private final MemberService memberService;
     private final SleepAdvisorService sleepAdvisorService;
     private final CharacterService characterService;
     private final MoneyService moneyService;
@@ -40,7 +40,7 @@ public class UserController {
 
 @PostMapping("/users/idCheck")
     public ResponseEntity<Boolean> idCheck(@RequestBody IdRequest idRequest){
-        String check= userService.idCheck(idRequest.getRequestId());
+        String check= memberService.idCheck(idRequest.getRequestId());
        if(check.equals("중복"))
        { return new ResponseEntity<>(false,HttpStatus.OK);}
        return new ResponseEntity<>(true,HttpStatus.FORBIDDEN);
@@ -48,7 +48,7 @@ public class UserController {
 
     @PostMapping("/users")
     public ResponseEntity<CommonResponse> signup(@RequestBody SignUpRequest signUpRequest) {
-        Member member = userService.join(signUpRequest);
+        Member member = memberService.join(signUpRequest);
         sleepAdvisorService.create(InitialSleepAdvisorDto.of(member,signUpRequest.getGoalSleepTime(),signUpRequest.getGoalWakeTime(),signUpRequest.getMinimumSleepTime()));
         characterService.createCharacter(InitialCharacterDto.of(member));
         moneyService.createMoney(InitialMoneyDto.of(member));
